fix: integer type management not consistent

+++ src/libs/computations/binary_decimal.go
@@ -5,13 +5,13 @@ import (
 )
 
 // Converts binary number to decimal.
-func BinaryToDecimal(binary int) int {
+func BinaryToDecimal(binary int64) int64 {
   binaryToArray := NumberArray(binary)
-  decimal := 0
+  decimal := int64(0)
 
   for i := 0; i < len(binaryToArray); i++ {
     element := binaryToArray[i]
-    decimal = decimal + (element * int(math.Pow(2, float64(i))))
+    decimal = decimal + (element * int64(math.Pow(2, float64(i))))
   }
 
   return decimal
