+++ src/libs/computations/number_array.go
@@ -1,7 +1,7 @@
 package computations
 
-func NumberArray(num int) []int {
-  var arrayOfNum []int
+func NumberArray(num int64) []int64 {
+  var arrayOfNum []int64
   
   for num != 0 {
     arrayOfNum = append(arrayOfNum, num % 10)
